Tweaked some things in search to give better results.
Direct match is determined by doing a sql query instead of using WordPress' get_page_by_title() function. This is more of what we want.

If your search is a direct match with a page title and there is only one result then we will redirect you right there. Otherwise we show the search results. This seems to be the right balance for now.